+++ tests/dummy/app/controllers/application.js
@@ -1,6 +1,12 @@
 import Controller from "@ember/controller";
+import { action } from "@ember/object";
 import { inject as service } from "@ember/service";
 
-export default Controller.extend({
-  session: service()
-});
+export default class ApplicationController extends Controller {
+  @service session;
+
+  @action
+  logout() {
+    this.session.invalidate();
+  }
+}
