--- tests/dummy/app/routes/logout.js
@@ -1,4 +1,0 @@
-import Route from "@ember/routing/route";
-import OIDCEndSessionRouteMixin from "ember-simple-auth-oidc/mixins/oidc-end-session-route-mixin";
-
-export default Route.extend(OIDCEndSessionRouteMixin, {});
